+++ phpmyfaq/src/phpMyFAQ/Configuration.php
@@ -31,12 +31,12 @@ class Configuration
     /**
      * @var array
      */
-    public $config = [];
+    public array $config = [];
 
     /**
      * @var string
      */
-    protected $tableName = 'faqconfig';
+    protected string $tableName = 'faqconfig';
 
     /**
      * Constructor.
@@ -88,7 +88,7 @@ class Configuration
      * @param mixed  $value
      * @return bool
      */
-    public function set(string $key, $value): bool
+    public function set(string $key, mixed $value): bool
     {
         $query = sprintf(
             "UPDATE %s%s SET config_value = '%s' WHERE config_name = '%s'",
@@ -196,7 +196,7 @@ class Configuration
     {
         $defaultUrl = $this->get('main.referenceURL');
 
-        if (substr($defaultUrl, -1) !== '/') {
+        if (!str_ends_with($defaultUrl, '/')) {
             return $defaultUrl . '/';
         } else {
             return $defaultUrl;
@@ -209,21 +209,18 @@ class Configuration
      * @param string $item Configuration item
      * @return mixed
      */
-    public function get(string $item)
+    public function get(string $item): mixed
     {
         if (!isset($this->config[$item])) {
             $this->getAll();
         }
 
         if (isset($this->config[$item])) {
-            switch ($this->config[$item]) {
-                case 'true':
-                    return true;
-                case 'false':
-                    return false;
-                default:
-                    return $this->config[$item];
-            }
+            return match ($this->config[$item]) {
+                'true' => true,
+                'false' => false,
+                default => $this->config[$item],
+            };
         }
 
         return null;
@@ -293,6 +290,62 @@ class Configuration
     }
 
     /**
+     * Sets the Active Directory configuration.
+     *
+     * @param string[] $ldapConfig
+     */
+    public function setActiveDirectoryConfig(array $ldapConfig): void
+    {
+        // Always add main Active Directory server
+        $this->config['core.activeDirectoryServer'][0] = [
+            'ad_server' => $ldapConfig['ad_server'],
+            'ad_port' => $ldapConfig['ad_port'],
+            'ad_user' => $ldapConfig['ad_user'],
+            'ad_password' => $ldapConfig['ad_password'],
+            'ad_base' => $ldapConfig['ad_base'],
+        ];
+
+        // Add multiple Active Directory servers if enabled
+        if (true === $this->get('ad.ad_use_multiple_servers')) {
+            $key = 1;
+            while ($key >= 1) {
+                if (isset($ldapConfig[$key])) {
+                    $this->config['core.activeDirectoryServer'][$key] = $ldapConfig[$key];
+                    ++$key;
+                } else {
+                    break;
+                }
+            }
+        }
+
+        // Set LDAP configuration
+        $this->config['core.ldapConfig'] = [
+            'ad_use_multiple_servers' => $this->get('ad.ad_use_multiple_servers'),
+            'ad_mapping' => $this->getActiveDirectoryMapping(),
+            'ad_use_domain_prefix' => $this->get('ad.ad_use_domain_prefix'),
+            'ad_options' => $this->getActiveDirectoryOptions(),
+            'ad_use_memberOf' => $this->get('ad.ad_use_memberOf'),
+            'ad_use_sasl' => $this->get('ad.ad_use_sasl'),
+            'ad_use_anonymous_login' => $this->get('ad.ad_use_anonymous_login'),
+        ];
+    }
+
+    /**
+     * Returns the Active Directory mapping configuration.
+     *
+     * @return string[]
+     */
+    public function getActiveDirectoryMapping(): array
+    {
+        return [
+            'name' => $this->get('ad.ad_mapping.name'),
+            'username' => $this->get('ad.ad_mapping.username'),
+            'mail' => $this->get('ad.ad_mapping.mail'),
+            'memberOf' => $this->get('ad.ad_mapping.memberOf')
+        ];
+    }
+
+    /**
      * Returns the LDAP mapping configuration.
      *
      * @return string[]
@@ -321,6 +374,19 @@ class Configuration
     }
 
     /**
+     * Returns the Active Directory options configuration.
+     *
+     * @return string[]
+     */
+    public function getActiveDirectoryOptions(): array
+    {
+        return [
+            'LDAP_OPT_PROTOCOL_VERSION' => $this->get('ad.ad_options.LDAP_OPT_PROTOCOL_VERSION'),
+            'LDAP_OPT_REFERRALS' => $this->get('ad.ad_options.LDAP_OPT_REFERRALS')
+        ];
+    }
+
+    /**
      * Returns the LDAP configuration.
      *
      * @return string[]
@@ -331,6 +397,16 @@ class Configuration
     }
 
     /**
+     * Returns the Active Directory configuration.
+     *
+     * @return string[]
+     */
+    public function getActiveDirectoryConfig(): array
+    {
+        return $this->config['core.activeDirectoryConfig'] ?? [];
+    }
+
+    /**
      * Returns the LDAP server(s).
      *
      * @return string[]
@@ -338,6 +414,16 @@ class Configuration
     public function getLdapServer(): array
     {
         return $this->config['core.ldapServer'] ?? [];
+    }
+
+    /**
+     * Returns the Active Directory server(s).
+     *
+     * @return string[]
+     */
+    public function getActiveDirectoryServer(): array
+    {
+        return $this->config['core.activeDirectoryServer'] ?? [];
     }
 
     /**
@@ -387,7 +473,7 @@ class Configuration
      * @param mixed  $value
      * @return bool|object
      */
-    public function add(string $name, $value)
+    public function add(string $name, mixed $value): object|bool
     {
         $insert = sprintf(
             "INSERT INTO
@@ -439,6 +525,8 @@ class Configuration
             'core.language', // Language
             'core.ldapServer', // Ldap
             'core.ldapConfig', // $LDAP
+            'core.activeDirectoryServer', // Active Directory
+            'core.activeDirectoryConfig', // $AD
             'core.elasticsearch', // Elasticsearch\Client
             'core.elasticsearchConfig' // $ES
         ];